requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.10.0
tqdm>=4.62.0
umap-learn>=0.5.0
plotly>=5.0.0
ipywidgets>=7.6.0